--- oidc-jersey-support/src/main/java/no/nav/security/token/support/core/jaxrs/servlet/JaxrsJwtTokenValidationFilter.java
@@ -1,13 +1,0 @@
-package no.nav.security.token.support.core.jaxrs.servlet;
-
-import no.nav.security.token.support.core.configuration.MultiIssuerConfiguration;
-import no.nav.security.token.support.core.filter.JwtTokenValidationFilter;
-import no.nav.security.token.support.core.jaxrs.JaxrsTokenValidationContextHolder;
-import no.nav.security.token.support.core.validation.JwtTokenValidationHandler;
-
-public class JaxrsJwtTokenValidationFilter extends JwtTokenValidationFilter {
-
-    public JaxrsJwtTokenValidationFilter(MultiIssuerConfiguration oidcConfig) {
-        super(new JwtTokenValidationHandler(oidcConfig), JaxrsTokenValidationContextHolder.getHolder());
-    }
-}
