+++ oidc-jersey-support/src/main/java/no/nav/security/token/support/jaxrs/servlet/JaxrsJwtTokenValidationFilter.java
@@ -1,0 +1,13 @@
+package no.nav.security.token.support.jaxrs.servlet;
+
+import no.nav.security.token.support.core.configuration.MultiIssuerConfiguration;
+import no.nav.security.token.support.filter.JwtTokenValidationFilter;
+import no.nav.security.token.support.jaxrs.JaxrsTokenValidationContextHolder;
+import no.nav.security.token.support.core.validation.JwtTokenValidationHandler;
+
+public class JaxrsJwtTokenValidationFilter extends JwtTokenValidationFilter {
+
+    public JaxrsJwtTokenValidationFilter(MultiIssuerConfiguration oidcConfig) {
+        super(new JwtTokenValidationHandler(oidcConfig), JaxrsTokenValidationContextHolder.getHolder());
+    }
+}
