+++ oidc-jersey-support/src/test/java/no/nav/security/token/support/jaxrs/rest/ProtectedClassResource.java
@@ -1,0 +1,18 @@
+package no.nav.security.token.support.jaxrs.rest;
+
+import no.nav.security.token.support.core.api.Protected;
+
+import javax.ws.rs.GET;
+import javax.ws.rs.Path;
+import javax.ws.rs.core.Response;
+
+@Path("class/protected")
+@Protected
+public class ProtectedClassResource {
+
+    @GET
+    public Response get() {
+        return Response.ok().build();
+    }
+
+}
