+++ oidc-spring-support-demo/src/main/java/no/nav/security/token/support/demo/spring/config/SecurityConfigurationDev.java
@@ -1,6 +1,6 @@
 package no.nav.security.token.support.demo.spring.config;
 
-import no.nav.security.token.support.core.test.support.spring.TokenGeneratorConfiguration;
+import no.nav.security.token.support.test.spring.TokenGeneratorConfiguration;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 import org.springframework.context.annotation.Profile;
