+++ oidc-spring-support/src/test/java/no/nav/security/token/support/spring/integrationtest/ProtectedRestControllerIntegrationTest.java
@@ -4,7 +4,7 @@ import com.nimbusds.jwt.JWT;
 import com.nimbusds.jwt.JWTClaimsSet;
 import com.nimbusds.jwt.PlainJWT;
 import io.restassured.module.mockmvc.RestAssuredMockMvc;
-import no.nav.security.token.support.core.test.support.JwkGenerator;
+import no.nav.security.token.support.test.JwkGenerator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,7 +24,7 @@ import java.util.concurrent.TimeUnit;
 
 import static io.restassured.module.mockmvc.RestAssuredMockMvc.given;
 import static no.nav.security.token.support.spring.integrationtest.ProtectedRestController.*;
-import static no.nav.security.token.support.core.test.support.JwtTokenGenerator.*;
+import static no.nav.security.token.support.test.JwtTokenGenerator.*;
 
 @SpringBootTest
 @ContextConfiguration(classes = {ProtectedApplication.class, ProtectedApplicationConfig.class})
